requirements.txt
requests>=2.31.0
django>=5.2.0
djangorestframework>=3.16.0
celery>=5.3.0
redis>=5.0.0
